+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.16.1] - 2024-07-18
+
+### Changed
+
+- Arktype updated to require `2.0.0-beta.0`, which should fix some typing issues with the adapter.
+
 ## [2.16.0] - 2024-07-09
 
 ### Added